requirements.txt
fastapi
langchain
langchain-openai
langgraph
llama-index>=0.10.20
numpy
openai
pandas
pdfkit
protobuf
PyMuPDF
python-dotenv
python-multipart
seaborn
tiktoken
tqdm
uvicorn



